backend/pyproject.toml
[tool.poetry]
name = "promvisio"
version = "0.1.0"
description = "Promvisio website backend"
authors = ["Satyuel <stockmann89@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
wagtail = "^5.0.1"
django-environ = "^0.10.0"
drf-spectacular = "^0.26.2"
sentry-sdk = {extras = ["django"], version = "^1.25.0"}
django-cors-headers = "^4.0.0"
psycopg = {extras = ["c"], version = "^3.1.9"}
wagtail-headless-preview = "^0.5.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.3.0"
pre-commit = "^3.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
